+++ core/exports/effects.ts
@@ -1,6 +1,0 @@
-export { config, denoConfig } from "$effects/config.ts";
-export { env } from "$effects/env.ts";
-export { hot } from "$effects/hot-update.ts";
-export { importmap } from "$effects/importmap.ts";
-export { io } from "$effects/io.ts";
-export { manifest } from "$effects/manifest.ts";