+++ core/src/plugins/manifest/manifest.ts
@@ -1,11 +1,11 @@
+import { hmr } from "$effects/hmr.ts";
+import { io } from "$effects/io.ts";
+import { manifest, manifestPath } from "$effects/manifest.ts";
+import { Handler, handlerFor } from "@radish/effect-system";
 import { assertExists } from "@std/assert";
 import { ensureDirSync, type ExpandGlobOptions } from "@std/fs";
 import { extname } from "@std/path";
 import { generatedFolder } from "../../constants.ts";
-import { Handler, handlerFor } from "@radish/effect-system";
-import { hot } from "$effects/hot-update.ts";
-import { io } from "$effects/io.ts";
-import { manifest, manifestPath } from "$effects/manifest.ts";
 import type { ManifestBase, Plugin } from "../../types.d.ts";
 import { expandGlobWorkspaceRelative } from "../../utils/fs.ts";
 import { extractImports } from "../../utils/parse.ts";
@@ -44,7 +44,7 @@ export const pluginManifest: Plugin = {
 
       return { entry, manifestObject };
     }),
-    handlerFor(hot.update, async ({ event, paths }) => {
+    handlerFor(hmr.update, async ({ event, paths }) => {
       if (event.isFile) {
         const manifestObject = await manifest.get();
         const manifestImports = manifestObject.imports;