+++ core/src/start.ts
@@ -3,7 +3,7 @@ import { env } from "$effects/env.ts";
 import { importmap } from "$effects/importmap.ts";
 import { manifest } from "$effects/manifest.ts";
 import * as effects from "@radish/effect-system";
-import { parseArgs } from "@std/cli/parse-args";
+import { parseArgs } from "@std/cli";
 import { UserAgent } from "@std/http";
 import {
   elementsFolder,
@@ -16,7 +16,7 @@ import { dev } from "./environment.ts";
 import { generateImportmap } from "./plugins/importmap/importmap.ts";
 import { updateManifest } from "./plugins/manifest/manifest.ts";
 import { createApp, type Handle } from "./server/app.ts";
-import type { CLIArgs, Config, ResolvedConfig } from "./types.d.ts";
+import type { CLIArgs, Config } from "./types.d.ts";
 
 const cliArgs: CLIArgs = Object.freeze(parseArgs(Deno.args, {
   boolean: ["dev", "env", "importmap", "manifest", "build", "server"],
@@ -41,11 +41,8 @@ export async function startApp(
   const handlers = config.plugins?.flatMap((plugin) => plugin.handlers) ?? [];
   using _ = new effects.HandlerScope(...handlers);
 
-  config = await configEffect.transform(config);
-
-  const resolvedConfig: ResolvedConfig = Object.freeze(
-    Object.assign({}, { args: cliArgs, ...config }),
-  );
+  config = await configEffect.transform({ ...config, args: cliArgs });
+  const resolvedConfig: Config = Object.freeze(config);
 
   effects.addHandlers([
     effects.handlerFor(configEffect.read, () => resolvedConfig),