+++ core/tests/shadowroot.test.ts
@@ -1,12 +1,11 @@
 import { fragments, serializeFragments } from "@radish/htmlcrunch";
-import { applyServerEffects } from "../src/walk.ts";
-import { assertEquals } from "@std/assert/equals";
+import { assertEquals } from "@std/assert";
 
 Deno.test("inline shadowroot", () => {
   const shadowRoot = serializeFragments(
     fragments.parseOrThrow(
       `<test-shadowroot></test-shadowroot>`,
-    ).map(applyServerEffects),
+    ),
   );
 
   assertEquals(