+++ Releases.md
@@ -1,57 +1,70 @@
 ### 2025.05.16
 
-#### @radish/effect-system 0.2.1 (patch) 
+#### @radish/effect-system 0.2.1 (patch)
+
 - chore(effect-system): export createState (#82)
 
 ### 2025.05.16
 
 #### @radish/core 1.0.0-alpha.39 (prerelease)
+
 - BREAKING(core): extract the effect system (#76)
 - chore(core): fix more imports (#74)
 - chore(core): fix rel import from @radish/runtime #73
 - chore(core): fix htmlcrunch rel import (#72)
 
 #### @radish/effect-system 0.1.0 (breaking)
-- BREAKING(effect-system): Snapshots, AsyncState and disposable HandlerScopes (#80)
+
+- BREAKING(effect-system): Snapshots, AsyncState and disposable HandlerScopes
+  (#80)
 - BREAKING(effect-system): disposable HandlerScope (#79)
 - chore(effect-system): add tests (#77)
 
 #### @radish/runtime 0.3.1 (patch)
+
 - chore(runtime): update package.json #75
 
 ### 2025.05.08
 
 #### @radish/core 1.0.0-alpha.38 (prerelease)
+
 - BREAKING(core): render effect (#65)
 - chore(core): update preview importmap (#70)
 
 #### @radish/htmlcrunch 1.2.0 (minor)
+
 - feat(htmlcrunch): node guard helpers
 - fix(htmlcrunch): fix attributes casing (#69)
 
 #### @radish/init 1.0.0-alpha.23 (prerelease)
+
 - chore(init): update init plugins (#67)
 
 #### @radish/runtime 0.3.0 (minor)
+
 - BREAKING(runtime): unprefixed directives (#68)
 
 ### 2025.04.30
 
 #### @radish/core 1.0.0-alpha.37 (prerelease)
+
 - feat(core): env effect (#55)
 - fix(core): slow types (#59)
 - chore(core): handle importmap insertion in the importmap plugin (#62)
 
 #### @radish/htmlcrunch 1.1.2 (patch)
+
 - fix(htmlcrunch): improve the parsing of custom element names (#58)
 
 #### @radish/init 1.0.0-alpha.22 (prerelease)
+
 - fix(init): update template script (#63)
 - chore(init): update template (#57)
 
 ### 2025.04.25
 
 #### @radish/core 1.0.0-alpha.36 (prerelease)
+
 - BREAKING(core): effect system (#41)
 - feat(core): support jsr modules on the client (#38)
 - chore(core): update preview importmap (#43)
@@ -59,10 +72,12 @@
 - chore(core): use asserts to explicit assumptions (#39)
 
 #### @radish/init 1.0.0-alpha.21 (prerelease)
+
 - chore(init): update template (#44)
 - chore(init): update templates (#42)
 
 ### 2025.03.26
 
 #### @radish/init 1.0.0-alpha.20 (prerelease)
+
 - refactor(init): idiomatic template fetching (#35)